moved the catalog fetching into catalog_tools
